+++ src/main/java/de/maxhenkel/replayvoicechat/mixin/GroupChatManagerMixin.java
@@ -1,7 +1,7 @@
 package de.maxhenkel.replayvoicechat.mixin;
 
 import de.maxhenkel.voicechat.voice.client.GroupChatManager;
-import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.gui.GuiGraphicsExtractor;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -12,7 +12,7 @@ import xyz.breadloaf.replaymodinterface.ReplayInterface;
 public class GroupChatManagerMixin {
 
     @Inject(method = "renderIcons", at = @At(value = "HEAD"), cancellable = true)
-    private static void renderIcons(GuiGraphics guiGraphics, CallbackInfo ci) {
+    private static void renderIcons(GuiGraphicsExtractor guiGraphics, CallbackInfo ci) {
         if (ReplayInterface.INSTANCE.isInReplayEditor) {
             ci.cancel();
         }